Extract es2k_port_manager from es2k_sde_target (#22)
- When the port management methods were separated from TdiSdeWrapper,
  the implementations were modified in place. This changelist moves the
  port methods from es2k_sde_target.cc to a new es2k_port_manager.cc
  file, paralleling the existing es2k_port_manager.h file.
